sdk: split query_verbs into module directory (3/9: extract overhead)
Move overhead, overhead_trim, and helpers to query_verbs/overhead.rs.
